requirements.txt
Adafruit-DHT==1.4.0
aliyun-python-sdk-core==2.13.25
aliyun-python-sdk-core-v3==2.13.11
aliyun-python-sdk-kms==2.11.0
autopep8==1.5.4
baidu-aip==2.2.18.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
crcmod==1.7
dlib==19.21.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna==2.10
jmespath==0.10.0
numpy==1.19.1
opencv-python==3.4.3.18
oss2==2.12.1
Pillow==7.2.0
PyAudio==0.2.11
pycodestyle==2.6.0
pycryptodome==3.9.8
PyJWT==1.7.1
PyMySQL==0.10.0
pytz==2020.1
requests==2.24.0
six==1.15.0
toml==0.10.1
twilio==6.44.2
urllib3==1.25.10
